srm_backend/pyproject.toml
[tool.poetry]
name = "srm-backend"
version = "0.1.0"
description = ""
authors = ["_Avvake_ <denisergo18@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.12"}
uvicorn = "^0.34.3"
pydantic-settings = "^2.9.1"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
alembic = "^1.16.1"
greenlet = "^3.2.3"
python-jose = "^3.5.0"
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
